+++ internal/format/builder.go
