+++ internal/output/stdout.go
