+++ internal/output/stdout_test.go
